#882 Create a shared project: Apply common ApiExceptionHandler for all modules
