+++ customer/src/it/resources/application.properties
@@ -16,3 +16,4 @@ keycloak.credentials.secret=NKAr3rnjwm9jlakgKpelukZGFaHYqIWE
 spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
+cors.allowed-origins=*
