+++ customer/src/main/resources/application.properties
@@ -48,3 +48,4 @@ resilience4j.circuitbreaker.instances.rest-circuit-breaker.failure-rate-threshol
 resilience4j.circuitbreaker.instances.rest-circuit-breaker.minimum-number-of-calls=5
 resilience4j.circuitbreaker.instances.rest-circuit-breaker.automatic-transition-from-open-to-half-open-enabled=true
 resilience4j.circuitbreaker.instances.rest-circuit-breaker.permitted-number-of-calls-in-half-open-state=3
+cors.allowed-origins=*
