+++ inventory/src/main/java/com/yas/inventory/InventoryApplication.java
@@ -1,12 +1,13 @@
 package com.yas.inventory;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import com.yas.inventory.config.ServiceUrlConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
-@EnableConfigurationProperties(ServiceUrlConfig.class)
+@SpringBootApplication(scanBasePackages = {"com.yas.inventory", "com.yas.commonlibrary"})
+@EnableConfigurationProperties({ServiceUrlConfig.class, CorsConfig.class})
 public class InventoryApplication {
     public static void main(String[] args) {
         SpringApplication.run(InventoryApplication.class, args);
