+++ inventory/src/test/resources/application.properties
@@ -10,4 +10,5 @@ spring.jpa.hibernate.ddl-auto=update
 spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
+cors.allowed-origins=*
