+++ location/src/main/java/com/yas/location/LocationApplication.java
@@ -1,9 +1,12 @@
 package com.yas.location;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"com.yas.location", "com.yas.commonlibrary"})
+@EnableConfigurationProperties(CorsConfig.class)
 public class LocationApplication {
 
     public static void main(String[] args) {
