+++ location/src/test/resources/application.properties
@@ -23,3 +23,4 @@ springdoc.packagesToScan=com.yas.location
 spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
+cors.allowed-origins=*
