+++ media/src/it/resources/application.properties
@@ -13,3 +13,4 @@ springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
 spring.jpa.open-in-view=false
 file.directory=images
+cors.allowed-origins=*
