+++ media/src/main/java/com/yas/media/MediaApplication.java
@@ -1,12 +1,13 @@
 package com.yas.media;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import com.yas.media.config.YasConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
-@EnableConfigurationProperties(YasConfig.class)
+@SpringBootApplication(scanBasePackages = {"com.yas.media", "com.yas.commonlibrary"})
+@EnableConfigurationProperties({YasConfig.class, CorsConfig.class})
 public class MediaApplication {
 
     public static void main(String[] args) {
