+++ media/src/main/resources/application.properties
@@ -26,3 +26,4 @@ springdoc.swagger-ui.oauth.client-id=swagger-ui
 springdoc.oauthflow.authorization-url=http://identity/realms/Yas/protocol/openid-connect/auth
 springdoc.oauthflow.token-url=http://identity/realms/Yas/protocol/openid-connect/token
 file.directory=/images/
+cors.allowed-origins=*
