+++ media/src/test/resources/application.properties
@@ -4,3 +4,4 @@ spring.datasource.username=sa
 spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=update
+cors.allowed-origins=*
