+++ payment-paypal/src/main/java/com/yas/paymentpaypal/PaymentPaypalApplication.java
@@ -1,12 +1,13 @@
 package com.yas.paymentpaypal;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import com.yas.paymentpaypal.config.ServiceUrlConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
-@EnableConfigurationProperties(ServiceUrlConfig.class)
+@SpringBootApplication(scanBasePackages = {"com.yas.paymentpaypal", "com.yas.commonlibrary"})
+@EnableConfigurationProperties({ServiceUrlConfig.class, CorsConfig.class})
 public class PaymentPaypalApplication {
 
     public static void main(String[] args) {
