--- payment-paypal/src/main/java/com/yas/paymentpaypal/config/CorsConfig.java
@@ -1,19 +1,0 @@
-package com.yas.paymentpaypal.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsConfig {
-    @Bean
-    public WebMvcConfigurer corsConfigure() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                registry.addMapping("/**").allowedMethods("*").allowedOrigins("*").allowedHeaders("*"); //NOSONAR
-            }
-        };
-    }
-}
