+++ payment-paypal/src/test/resources/application.properties
@@ -18,3 +18,4 @@ spring.security.oauth2.resourceserver.jwt.issuer-uri=http://identity/realms/Yas
 springdoc.oauthflow.authorization-url=http://identity/realms/Yas/protocol/openid-connect/auth
 springdoc.oauthflow.token-url=http://identity/realms/Yas/protocol/openid-connect/token
 spring.jpa.open-in-view=false
+cors.allowed-origins=*
