+++ payment/src/it/resources/application.properties
@@ -12,3 +12,4 @@ spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
 spring.jpa.open-in-view=false
+cors.allowed-origins=*
