+++ payment/src/main/java/com/yas/payment/PaymentApplication.java
@@ -1,12 +1,13 @@
 package com.yas.payment;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import com.yas.payment.config.ServiceUrlConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
-@EnableConfigurationProperties(ServiceUrlConfig.class)
+@SpringBootApplication(scanBasePackages = {"com.yas.payment", "com.yas.commonlibrary"})
+@EnableConfigurationProperties({ServiceUrlConfig.class, CorsConfig.class})
 public class PaymentApplication {
 
     public static void main(String[] args) {
