+++ payment/src/test/resources/application.properties
@@ -26,3 +26,4 @@ springdoc.api-docs.enabled=true
 # swagger-ui custom path
 springdoc.swagger-ui.path=/swagger-ui.html
 springdoc.packagesToScan=com.yas.payment
+cors.allowed-origins=*
