+++ promotion/src/main/java/com/yas/promotion/PromotionApplication.java
@@ -1,12 +1,13 @@
 package com.yas.promotion;
 
+import com.yas.commonlibrary.config.CorsConfig;
 import com.yas.promotion.config.ServiceUrlConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
-@SpringBootApplication
-@EnableConfigurationProperties(ServiceUrlConfig.class)
+@SpringBootApplication(scanBasePackages = {"com.yas.promotion", "com.yas.commonlibrary"})
+@EnableConfigurationProperties({ServiceUrlConfig.class, CorsConfig.class})
 public class PromotionApplication {
 
     public static void main(String[] args) {
