+++ promotion/src/main/resources/application.properties
@@ -36,3 +36,4 @@ springdoc.swagger-ui.oauth.use-pkce-with-authorization-code-grant=true
 springdoc.swagger-ui.oauth.client-id=swagger-ui
 springdoc.oauthflow.authorization-url=http://identity/realms/Yas/protocol/openid-connect/auth
 springdoc.oauthflow.token-url=http://identity/realms/Yas/protocol/openid-connect/token
+cors.allowed-origins=*
