+++ promotion/src/test/resources/application.properties
@@ -16,5 +16,6 @@ spring.liquibase.enabled=false
 spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
+cors.allowed-origins=*
