+++ rating/src/it/resources/application.properties
@@ -13,5 +13,6 @@ spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
 spring.jpa.open-in-view=false
+cors.allowed-origins=*
 
 cors.allowed-origins=*
