+++ sampledata/src/main/resources/application.properties
@@ -38,5 +38,6 @@ spring.jpa.properties.hibernate.boot.allow_jdbc_metadata_access=false
 # Database initialization should typically be performed outside of Spring lifecycle
 spring.jpa.hibernate.ddl-auto=none
 spring.sql.init.mode=never
+cors.allowed-origins=*
 
 cors.allowed-origins=*
