+++ sampledata/src/test/resources/application.properties
@@ -13,5 +13,4 @@ spring.profiles.active=test
 spring.security.oauth2.resourceserver.jwt.issuer-uri=test
 springdoc.oauthflow.authorization-url=test
 springdoc.oauthflow.token-url=test
-
 cors.allowed-origins=*
